requirements.txt
# Ritual File Renamer - Python Dependencies

# OCR
pytesseract>=0.3.10

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Video Processing (uses system ffmpeg)
# ffmpeg-python>=0.2.0  # Optional: Python bindings for ffmpeg

# File Operations
pathlib2>=2.3.7; python_version < "3.4"

# Metadata
exifread>=3.0.0  # EXIF reading for original capture time

# GUI
customtkinter>=5.2.0  # Modern tkinter UI

# Utilities
tqdm>=4.66.0  # Progress bar
